+++ proposal_test.go
@@ -95,7 +95,7 @@ func Test_BaseProposal_SetChainMetadata(t *testing.T) {
 	})
 
 	assert.Equal(t, uint64(0), proposal.ChainMetadata[chaintest.Chain1Selector].StartingOpCount)
-	assert.Equal(t, "", proposal.ChainMetadata[chaintest.Chain1Selector].MCMAddress)
+	assert.Empty(t, proposal.ChainMetadata[chaintest.Chain1Selector].MCMAddress)
 }
 
 func Test_NewProposal(t *testing.T) {
@@ -252,8 +252,7 @@ func Test_WriteProposal(t *testing.T) {
 				"chainMetadata": {
 					"3379446385462418246": {
 						"mcmAddress": "",
-						"startingOpCount": 0,
-						"additionalFields": null
+						"startingOpCount": 0
 					}
 				},
 				"operations": [